+++ src/SettingsList.h
@@ -562,6 +562,12 @@ inline const std::vector<SettingInfo>& getBaseSettingsList() {
     add(buildSleepScreenSetting());
     add(SettingInfo::Enum(StrId::STR_SLEEP_COVER_MODE, &CrossPointSettings::sleepScreenCoverMode,
                           {StrId::STR_FIT, StrId::STR_CROP}, "sleepScreenCoverMode", StrId::STR_CAT_DISPLAY));
+    add(SettingInfo::Value(
+        StrId::STR_SLIDESHOW_INTERVAL, &CrossPointSettings::slideshowIntervalMinutes,
+        {CrossPointSettings::MIN_SLIDESHOW_INTERVAL_MINUTES, CrossPointSettings::MAX_SLIDESHOW_INTERVAL_MINUTES, 1},
+        "slideshowIntervalMinutes", StrId::STR_CAT_DISPLAY));
+    add(SettingInfo::Enum(StrId::STR_SLIDESHOW_DISPLAY_MODE, &CrossPointSettings::slideshowScaleMode,
+                          {StrId::STR_FIT, StrId::STR_CROP}, "slideshowScaleMode", StrId::STR_CAT_DISPLAY));
     add(SettingInfo::Enum(StrId::STR_SLEEP_COVER_FILTER, &CrossPointSettings::sleepScreenCoverFilter,
                           {StrId::STR_NONE_OPT, StrId::STR_FILTER_CONTRAST, StrId::STR_INVERTED},
                           "sleepScreenCoverFilter", StrId::STR_CAT_DISPLAY));